requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
openai>=1.12.0
pypdf==4.0.1
python-dotenv==1.0.1
Werkzeug==3.0.1
psycopg2-binary==2.9.9
